+++ src/zope/hookable/__init__.py
@@ -13,6 +13,12 @@
 ##############################################################################
 """Hookable object support
 """
+import os
+import platform
+
+
+_PYPY = platform.python_implementation() in ('PyPy', 'Jython')
+_PURE_PYTHON = os.environ.get('PURE_PYTHON', _PYPY)
 
 class _py_hookable(object):
     __slots__ = ('_original', '_implementation')
@@ -56,9 +62,12 @@ class _py_hookable(object):
     def __call__(self, *args, **kw):
         return self._implementation(*args, **kw)
 
-hookable = _py_hookable
-
 try:
-    from ._zope_hookable import hookable
+    from zope.hookable._zope_hookable import hookable as _c_hookable
 except ImportError: # pragma: no cover
-    pass
+    _c_hookable = None
+
+if _PURE_PYTHON or _c_hookable is None: # pragma: no cover
+    hookable = _py_hookable
+else:
+    hookable = _c_hookable